requirements.txt
asgiref==3.7.2
Django==4.2.3
django-filter==23.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
